Refactoring and more comments for codegen
