+++ src/core/xc_fallback/README.md
@@ -1,0 +1,5 @@
+This folder contains fallback implementations of commonly used exchange-correlation
+functionals, which can be used with any Array or floating point type.
+
+The implementations are based upon the Julia code generated using
+the scripts from DFTK.jl/codegen
